Migrates to DataComponents for NeoForge 1.21.1
Addresses compatibility issues with NeoForge 1.21.1 by migrating from NBT data to DataComponents.

This involves updating several classes to use LuaNbtAccessor for handling item data. The partial tick value is now correctly passed in the BeforeRenderHandEvent and used consistently in rendering logic.

Temporarily disables some features like Ammo/Block index retrieval due to data structure changes, these will be addressed in subsequent commits.

+++ src/main/java/com/tacz/guns/client/event/FirstPersonRenderEvent.java
@@ -53,11 +53,11 @@ public class FirstPersonRenderEvent {
             }
             boolean flag = ItemStack.matches(player.getMainHandItem(), stack);
             if (flag && renderer.needReInit(stack)) {
-                renderer.tryInit(stack, player, event.getPartialTick());
+                renderer.tryInit(stack, player, (float) event.getPartialTick());
             }
 
             renderer.renderFirstPerson(player, stack, transformType, event.getPoseStack(), event.getMultiBufferSource(),
-                    event.getPackedLight(), event.getPartialTick());
+                    event.getPackedLight(), (float) event.getPartialTick());
             event.setCanceled(true);
         }
     }
